refactor(api): refactor create user handler with auth service
